// www/js/GroupForums/ts/react/shared/types.ts
interface User {
  hasVerifiedBadge: boolean;
  userId: number;
  username: string;
  displayName: string;
}
// ...
export type RoleColors = typeof RoleColorValues[keyof typeof RoleColorValues];

interface Role {
  id: number;
  name: string;
  rank: number;
  memberCount?: number;
  description?: string;
  isBase?: boolean;
  color?: RoleColors;
  isPrivate?: boolean;
}

interface Member {
  user: User;
  role: Role;
}

interface GroupPermissions {
  groupPostsPermissions: {
    viewWall: boolean;
    postToWall: boolean;
    deleteFromWall: boolean;
    viewStatus: boolean;
    postToStatus: boolean;
  };
  groupMembershipPermissions: {
    changeRank: boolean;
    inviteMembers: boolean;
    removeMembers: boolean;
    banMembers: boolean;
  };
  groupManagementPermissions: {
    manageRelationships: boolean;
    manageClan: boolean;
    viewAuditLogs: boolean;
  };
  groupEconomyPermissions: {
    spendGroupFunds: boolean;
    advertiseGroup: boolean;
    createItems: boolean;
    manageItems: boolean;
    addGroupPlaces: boolean;
    manageGroupGames: boolean;
    viewGroupPayouts: boolean;
    viewAnalytics: boolean;
  };
  groupOpenCloudPermissions: {
    useCloudAuthentication: boolean;
    administerCloudAuthentication: boolean;
  };
  groupForumsPermissions: {
    manageCategories: boolean;
    createPosts: boolean;
    removePosts: boolean;
    lockPosts: boolean;
    pinPosts: boolean;
    createComments: boolean;
    removeComments: boolean;
    createBugReports?: boolean;
  };
  groupContentModerationPermissions: {
    manageKeywordBlockList: boolean;
    viewKeywordBlockList: boolean;
  };
}

interface GroupChannelPermissions {
  channelId: string;
  groupForumsPermissions: {
    manageCategories: boolean;
    createPosts: boolean;
    removePosts: boolean;
    lockPosts: boolean;
    pinPosts: boolean;
    createComments: boolean;
    removeComments: boolean;
    createBugReports?: boolean;
  };
}

interface Group {
  permissions?: GroupPermissions;
  id: number;
  name: string;
  publicEntryAllowed?: boolean;
  owner?: User;
  role?: Role;
  description?: string;
  memberCount?: number;
  previousName?: string;
  hasVerifiedBadge?: boolean;
}
// ...
interface GroupConfigurationMetadata {
  roleConfiguration: {
    cost: number;
    descriptionMaxLength: number;
    limit: number;
    maxRank: number;
    minRank: number;
    nameMaxLength: number;
  };
  groupConfiguration: {
    nameMaxLength: number;
    descriptionMaxLength: number;
    iconMaxFileSizeMb: number;
    coverPhotoMaxFileSizeMb: number;
    validCoverPhotoDimensions: string;
    cost: number;
  };
  groupNameChangeConfiguration: {
    cost: number;
    cooldownInDays: number;
    ownershipCooldownInDays: number;
  };
}